*** empty log message ***
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@638 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/run-time/sig.c
@@ -740,7 +740,7 @@ int sig;
 	int i;
 
 #ifdef HAS_SYS_SIGLIST
-	if (sig >= 0 && sig < NSIG && 0 < strlen(sys_siglist[sig]))
+	if (sig >= 0 && sig < NSIG && 0 < (unsigned int)strlen(sys_siglist[sig]))
 		return (char *) sys_siglist[sig];
 #endif
